Fix all clang-tidy and ruff formatting warnings
- Narrowing conversions: add static_cast where needed
- Uppercase float suffixes: 0.0f → 0.0F
- Use std::max instead of manual comparison
- Use .empty() instead of .size() checks
- Remove dead stores (unused variables)
- Split multi-variable declarations
- Refactor nested ternary into if/else
- Fix ruff formatting on type: ignore comments

+++ src/pjrt_plugin/ops/convolution_ops.mm
@@ -202,7 +202,7 @@ static ProcessResult HandleConvolution(HandlerContext& ctx) {
 
     // Determine spatial rank and batch group count
     size_t spatialRank = inputSpatialDims.size();
-    int64_t batchGroupCount = convOp.getBatchGroupCount();
+    int64_t batchGroupCount = static_cast<int64_t>(convOp.getBatchGroupCount());
 
     bool is1D = (spatialRank == 1);
     if (spatialRank > 2) {
